Refactor normalizer
Add metrics to normalizer service without exporting its internals.
